pyproject.toml
[project]
name = "salesforce-netsuite-mcp"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.9,<2",
    "simple-salesforce>=1.12,<2",
    "httpx>=0.27,<1",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "PyJWT>=2.8,<3",
    "cryptography>=41,<44",
    "requests>=2.31,<3",
    "python-dotenv>=1.0,<2",
    "uvicorn>=0.30,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "ruff>=0.4,<1",
]
